Review/cleanup/refactor
- updated README, other docs to reflect recent changes
- addressed a number of TODO items that were listed in the code
- implemented multi-cluster testing
- reorganized issue-related reproducer tests
- reorganized other tests' location and t.Run names
- added skipFunction to policyTestCase
- removed dead code
