+++ test/integration/acceptance/custom/hello_policy/hostnames_test.go
@@ -7,7 +7,6 @@ package hello_policy
 // - cross testing (claim on router and vice versa)
 // - full setup checking (create service and expose; check they appear/disappear; perhaps even curl the service).
 // - Add also different removals and reinstates of policy (actual removal, changed namespace list)
-// - Change it to use a secondary cluster, if available
 
 import (
 	"fmt"
