+++ test/integration/acceptance/custom/hello_policy/link_test.go
@@ -343,116 +343,3 @@ func testLinkPolicy(t *testing.T, pub, prv *base.ClusterContext) {
 	}.run(t, pub, prv)
 
 }
-
-// Testing for #789
-//
-// https://github.com/skupperproject/skupper/issues/789
-//
-// It's basically the same previously-created-token test case above, but on a
-// loop.
-func testLinkIssue789(t *testing.T, pub, prv *base.ClusterContext) {
-
-	base.SkipIssueTests(t)
-
-	testTable := []policyTestCase{
-		{
-			name: "init",
-			steps: []policyTestStep{
-				{
-					name:     "execute",
-					parallel: true,
-					cliScenarios: []cli.TestScenario{
-						skupperInitInteriorTestScenario(pub, "", true),
-						skupperInitEdgeTestScenario(prv, "", true),
-					},
-				},
-			},
-		}, {
-			name: "previously-created-token",
-			steps: []policyTestStep{
-				{
-					name: "prepare",
-					pubPolicy: []skupperv1.SkupperClusterPolicySpec{
-						allowIncomingLinkPolicy(pub.Namespace, true),
-					},
-					getChecks: []policyGetCheck{
-						{
-							allowIncoming: cli.Boolp(true),
-							cluster:       pub,
-						}, {
-							allowIncoming: cli.Boolp(false),
-							allowedHosts:  []string{"any"},
-							cluster:       prv,
-						},
-					},
-					cliScenarios: []cli.TestScenario{
-						createTokenPolicyScenario(pub, "", testPath, "previous", true),
-					},
-				},
-			},
-		},
-	}
-
-	for i := 0; i < 30; i++ {
-		testTable = append(testTable, policyTestCase{
-			name: "run",
-			steps: []policyTestStep{
-				{
-					name: "disallow-and-create-link",
-					pubPolicy: []skupperv1.SkupperClusterPolicySpec{
-						allowIncomingLinkPolicy(pub.Namespace, false),
-					},
-					getChecks: []policyGetCheck{
-						{
-							cluster:       pub,
-							allowIncoming: cli.Boolp(false),
-						},
-					},
-					cliScenarios: []cli.TestScenario{
-						createLinkTestScenario(prv, "", "previous", false),
-						linkStatusTestScenario(prv, "", "previous", false),
-					},
-				}, {
-					name: "re-allow-and-check-link",
-					pubPolicy: []skupperv1.SkupperClusterPolicySpec{
-						allowIncomingLinkPolicy(pub.Namespace, true),
-					},
-					getChecks: []policyGetCheck{
-						{
-							cluster:       pub,
-							allowIncoming: cli.Boolp(true),
-						},
-					},
-					cliScenarios: []cli.TestScenario{
-						linkStatusTestScenario(prv, "now", "previous", true),
-						sitesConnectedTestScenario(pub, prv, "", "previous"),
-						linkDeleteTestScenario(prv, "", "previous"),
-					},
-				},
-			},
-		})
-	}
-
-	testTable = append(testTable, policyTestCase{
-		name: "cleanup",
-		steps: []policyTestStep{
-			{
-				name:     "delete",
-				parallel: true,
-				cliScenarios: []cli.TestScenario{
-					deleteSkupperTestScenario(pub, "pub"),
-					deleteSkupperTestScenario(prv, "prv"),
-				},
-			},
-		},
-	})
-
-	policyTestRunner{
-		testCases:    testTable,
-		keepPolicies: true,
-		prvPolicies: []skupperv1.SkupperClusterPolicySpec{
-			allowedOutgoingLinksHostnamesPolicy(prv.Namespace, []string{"*"}),
-		},
-	}.run(t, pub, prv)
-
-}
