+++ test/integration/acceptance/custom/hello_policy/service_test.go
@@ -5,7 +5,6 @@ package hello_policy
 
 import (
 	"testing"
-	"time"
 
 	"github.com/skupperproject/skupper/api/types"
 	"github.com/skupperproject/skupper/pkg/apis/skupper/v1alpha1"
@@ -810,13 +809,6 @@ func testServicePolicy(t *testing.T, pub, prv *base.ClusterContext) {
 						serviceCheckFrontStatusTestScenario(pub, "", []string{"hello-world-backend", "hello-world-backend"}, []string{}, []string{}),
 						serviceCheckBackStatusTestScenario(prv, "", []string{"hello-world-backend"}, []string{}, []string{"hello-world-backend"}),
 					},
-				}, {
-					name: "todo-skip",
-					skip: func() string {
-						// Note that it is services-there--but-not-bound that needs fixing; this one is here just to document, and
-						// should be removed once that one is fixed.  TODO
-						return "TODO: services.StatusTester needs refactored for proper non-bound check"
-					},
 				},
 			},
 		}, {
@@ -843,359 +835,6 @@ func testServicePolicy(t *testing.T, pub, prv *base.ClusterContext) {
 				AllowIncomingLinks:            true,
 				AllowedOutgoingLinksHostnames: []string{"*"},
 				AllowedExposedResources:       []string{"*"},
-			},
-		},
-		// Add background policies; policies that are not removed across
-		// runs
-	}.run(t, pub, prv)
-
-}
-
-// If a service is up and running, then we remove all policies and immediatelly reallow them: will
-// the services be removed?
-//
-// The result may depend on cluster load and other factors, so we run it multiple times (runs+1)
-//
-// The test removes the policy in two different ways: one is actual removal, the other is reusing the
-// policy, but removing only the target namespace from its list
-//
-// TODO: test with changing the setting on AllowedServices, instead?
-//
-// This test expects the service to be removed.  See removeReallowKeep for the alternative.  Only
-// one of them should be part of the test set, but which is unclear.
-func removeReallowRemove(pub, prv *base.ClusterContext, runs int) (allTestSteps []policyTestStep) {
-
-	baseTestSteps := []policyTestStep{
-		{
-			name:     "add-policy--check-services-absent",
-			parallel: true,
-			pubPolicy: []v1alpha1.SkupperClusterPolicySpec{
-				{
-					Namespaces:      []string{pub.Namespace},
-					AllowedServices: []string{"*"},
-				}, {
-					Namespaces:      []string{prv.Namespace},
-					AllowedServices: []string{"*"},
-				},
-			},
-			cliScenarios: []cli.TestScenario{
-				serviceCheckAbsentTestScenario(pub, "frontend", []string{"hello-world-frontend", "hello-world-backend"}),
-				serviceCheckAbsentTestScenario(prv, "backend", []string{"hello-world-frontend", "hello-world-backend"}),
-			},
-		}, {
-			name:     "create-service",
-			parallel: true,
-			cliScenarios: []cli.TestScenario{
-				serviceCreateFrontTestScenario(pub, "", true),
-				serviceCreateBackTestScenario(prv, "", true),
-			},
-		}, {
-			name:     "check-services-created",
-			parallel: true,
-			cliScenarios: []cli.TestScenario{
-				serviceCheckFrontStatusTestScenario(pub, "", []string{"hello-world-frontend", "hello-world-backend"}, []string{}, []string{}),
-				serviceCheckBackStatusTestScenario(prv, "", []string{"hello-world-frontend", "hello-world-backend"}, []string{}, []string{}),
-			},
-		}, {
-			name: "remove-policy",
-			pubPolicy: []v1alpha1.SkupperClusterPolicySpec{
-				{
-					Namespaces:      []string{"non-existing-namespace"},
-					AllowedServices: []string{"*"},
-				}, {
-					Namespaces: []string{"REMOVE"},
-				},
-			},
-			// We do not run anything on this policy, as we want to the next step (recreating the
-			// policy) to be executed right away
-		},
-	}
-
-	// The first execution is just preparation; we need to run at the very least two full cycles
-	for i := 0; i < runs+1; i++ {
-		allTestSteps = append(allTestSteps, baseTestSteps...)
-	}
-
-	return
-}
-
-// If a service is up and running, then we remove all policies and immediatelly reallow them: will
-// the services be removed?
-//
-// The result may depend on cluster load and other factors, so we run it multiple times (runs+1)
-//
-// The test removes the policy in two different ways: one is actual removal, the other is reusing the
-// policy, but removing only the target namespace from its list
-//
-// TODO:
-// - test with changing the setting on AllowedServices, instead
-// - test with switching the policy: remove existing policy but add another that allows
-//
-// This test expects the service to be kept.  See removeReallowKeep for the alternative.  Only
-// one of them should be part of the test set, but which is unclear.
-func removeReallowKeep(pub, prv *base.ClusterContext, runs int) (allTestSteps []policyTestStep) {
-
-	// TODO: IDEA: add an ever increasing sleep between removal and recreation; check
-	// at what point it changes
-
-	allTestSteps = append(allTestSteps, policyTestStep{
-		name:     "initial-service-creation",
-		parallel: true,
-		pubPolicy: []v1alpha1.SkupperClusterPolicySpec{
-			{
-				Namespaces:      []string{pub.Namespace},
-				AllowedServices: []string{"*"},
-			}, {
-				Namespaces:      []string{prv.Namespace},
-				AllowedServices: []string{"*"},
-			},
-		},
-		cliScenarios: []cli.TestScenario{
-			serviceCreateFrontTestScenario(pub, "", true),
-			serviceCreateBackTestScenario(prv, "", true),
-		},
-	})
-
-	baseTestSteps := []policyTestStep{
-		{
-			name:     "add-policy--check-services-present",
-			parallel: true,
-			pubPolicy: []v1alpha1.SkupperClusterPolicySpec{
-				{
-					Namespaces:      []string{pub.Namespace},
-					AllowedServices: []string{"*"},
-				}, {
-					Namespaces:      []string{prv.Namespace},
-					AllowedServices: []string{"*"},
-				},
-			},
-			cliScenarios: []cli.TestScenario{
-				serviceCheckFrontStatusTestScenario(pub, "", []string{"hello-world-frontend", "hello-world-backend"}, []string{}, []string{}),
-				serviceCheckBackStatusTestScenario(prv, "", []string{"hello-world-frontend", "hello-world-backend"}, []string{}, []string{}),
-			},
-		}, {
-			name: "remove-policy",
-			pubPolicy: []v1alpha1.SkupperClusterPolicySpec{
-				{
-					Namespaces:      []string{"non-existing-namespace"},
-					AllowedServices: []string{"*"},
-				}, {
-					Namespaces: []string{"REMOVE"},
-				},
-			},
-			// We do not run anything on this policy, as we want to the next step (recreating the
-			// policy) to be executed right away
-		},
-	}
-
-	// The first execution is just preparation; we need to run at the very least two full cycles
-	for i := 0; i < runs+1; i++ {
-		s := baseTestSteps
-		s[1].sleep = time.Duration(5000/runs*i) * time.Millisecond
-		allTestSteps = append(allTestSteps, s...)
-	}
-
-	return
-}
-
-// Can a service be created immediatelly after a policy allowing it?  If not, do we get a
-// proper error message for that?
-func allowAndCreate(pub, prv *base.ClusterContext, runs int) (allTestSteps []policyTestStep) {
-
-	allTestSteps = append(allTestSteps, policyTestStep{
-		// These policies are created just so the first cycle of the
-		// baseTestSteps has something to remove
-		name: "initial-dummy-policies",
-		pubPolicy: []v1alpha1.SkupperClusterPolicySpec{
-			{
-				Namespaces:      []string{pub.Namespace},
-				AllowedServices: []string{"*"},
-			}, {
-				Namespaces:      []string{prv.Namespace},
-				AllowedServices: []string{"*"},
-			},
-		},
-	})
-
-	baseTestSteps := []policyTestStep{
-		{
-			name:     "remove-policy--check-services-absent",
-			parallel: true,
-			pubPolicy: []v1alpha1.SkupperClusterPolicySpec{
-				{
-					Namespaces:      []string{"non-existing"},
-					AllowedServices: []string{"*"},
-				}, {
-					Namespaces: []string{"REMOVE"},
-				},
-			},
-			cliScenarios: []cli.TestScenario{
-				serviceCheckAbsentTestScenario(pub, "frontend", []string{"hello-world-frontend", "hello-world-backend"}),
-				serviceCheckAbsentTestScenario(prv, "backend", []string{"hello-world-frontend", "hello-world-backend"}),
-			},
-		}, {
-			name: "add-policy--create-service",
-			pubPolicy: []v1alpha1.SkupperClusterPolicySpec{
-				{
-					Namespaces:      []string{pub.Namespace},
-					AllowedServices: []string{"*"},
-				}, {
-					Namespaces:      []string{prv.Namespace},
-					AllowedServices: []string{"*"},
-				},
-			},
-			// We try to add services right after giving policy permission.  Should that work, should that fail?
-			// If it fails, it should give a failure response.
-			parallel: true,
-			cliScenarios: []cli.TestScenario{
-				serviceCreateFrontTestScenario(pub, "", true),
-				serviceCreateBackTestScenario(prv, "", true),
-			},
-		}, {
-			// This is the crux of this testing.  If we got to this point, the service should have been created
-			name:     "check-services-created",
-			parallel: true,
-			cliScenarios: []cli.TestScenario{
-				serviceCheckFrontStatusTestScenario(pub, "", []string{"hello-world-frontend", "hello-world-backend"}, []string{}, []string{}),
-				serviceCheckBackStatusTestScenario(prv, "", []string{"hello-world-frontend", "hello-world-backend"}, []string{}, []string{}),
-			},
-		},
-	}
-
-	// The first execution is just preparation; we need to run at the very least two full cycles
-	for i := 0; i < runs+1; i++ {
-		s := baseTestSteps
-		// s[1].sleep = time.Duration(5000/runs*i) * time.Millisecond
-		allTestSteps = append(allTestSteps, s...)
-	}
-
-	return
-
-}
-
-// As the name says: if the last policy that allows a service to exist is removed, the
-// service needs to be removed by the system
-func removePolicyRemoveServices(pub, prv *base.ClusterContext, runs int) (allTestSteps []policyTestStep) {
-
-	baseTestSteps := []policyTestStep{
-		{
-			// We're just adding the policy, so there should be no services
-			// around.  Neither at start of test, nor at start of cycle
-			// (that's what we're testing, that policy removal got rid of
-			// running services)
-			name: "add-policy--check-services-absent",
-			pubPolicy: []v1alpha1.SkupperClusterPolicySpec{
-				{
-					Namespaces:      []string{pub.Namespace},
-					AllowedServices: []string{"*"},
-				}, {
-					Namespaces:      []string{prv.Namespace},
-					AllowedServices: []string{"*"},
-				},
-			},
-			// We wait a few seconds to ensure we do not fall on
-			// https://github.com/skupperproject/skupper/issues/728
-			sleep: time.Duration(10) * time.Second,
-		}, {
-			name:     "create-services",
-			parallel: true,
-			cliScenarios: []cli.TestScenario{
-				serviceCreateFrontTestScenario(pub, "", true),
-				serviceCreateBackTestScenario(prv, "", true),
-			},
-		}, {
-			name:     "check-services",
-			parallel: true,
-			cliScenarios: []cli.TestScenario{
-				serviceCheckFrontStatusTestScenario(pub, "", []string{"hello-world-frontend", "hello-world-backend"}, []string{}, []string{}),
-				serviceCheckBackStatusTestScenario(prv, "", []string{"hello-world-frontend", "hello-world-backend"}, []string{}, []string{}),
-			},
-		}, {
-			// This is the crux of this testing.  If we remove policies, the services
-			// must be removed
-			name:     "remove-policy--check-services-removed",
-			parallel: true,
-			pubPolicy: []v1alpha1.SkupperClusterPolicySpec{
-				{
-					Namespaces:      []string{"non-existing"},
-					AllowedServices: []string{"*"},
-				}, {
-					Namespaces: []string{"REMOVE"},
-				},
-			},
-			cliScenarios: []cli.TestScenario{
-				serviceCheckAbsentTestScenario(pub, "frontend", []string{"hello-world-frontend", "hello-world-backend"}),
-				serviceCheckAbsentTestScenario(prv, "backend", []string{"hello-world-frontend", "hello-world-backend"}),
-			},
-		},
-	}
-
-	// The first execution is just preparation; we need to run at the very least two full cycles
-	for i := 0; i < runs+1; i++ {
-		s := baseTestSteps
-		// s[1].sleep = time.Duration(5000/runs*i) * time.Millisecond
-		allTestSteps = append(allTestSteps, s...)
-	}
-
-	return
-
-}
-
-// This is a good candidate to remove on t.Short(), or to skip by default
-func testServicePolicyTransitions(t *testing.T, pub, prv *base.ClusterContext) {
-
-	testTable := []policyTestCase{
-		{
-			name: "initialize",
-			steps: []policyTestStep{
-				{
-					name:     "skupper-init",
-					parallel: true,
-					cliScenarios: []cli.TestScenario{
-						skupperInitInteriorTestScenario(pub, "", true),
-						skupperInitEdgeTestScenario(prv, "", true),
-					},
-				}, {
-					name: "connect",
-					cliScenarios: []cli.TestScenario{
-						connectSitesTestScenario(pub, prv, "", "service"),
-					},
-				},
-			},
-			//		}, {
-			//			// This is testing for https://github.com/skupperproject/skupper/issues/728
-			//			name:  "allow-policy--and--immediatelly-create-service",
-			//			steps: allowAndCreate(pub, prv, 100),
-			//		}, {
-			//			// This is testing for https://github.com/skupperproject/skupper/issues/727
-			//			name:  "remove-policy-reallow--check-service-removed",
-			////			steps: removeReallowKeep(pub, prv, 500),
-			//		}, {
-			//			// This is testing for ???
-			//			name:  "remove-policy--remove-service",
-			//			steps: removePolicyRemoveServices(pub, prv, 400),
-		}, {
-			name: "cleanup",
-			steps: []policyTestStep{
-				{
-					name:     "delete",
-					parallel: true,
-					cliScenarios: []cli.TestScenario{
-						deleteSkupperTestScenario(pub, "pub"),
-						deleteSkupperTestScenario(prv, "prv"),
-					},
-				},
-			},
-		},
-	}
-
-	policyTestRunner{
-		testCases: testTable,
-		pubPolicies: []v1alpha1.SkupperClusterPolicySpec{
-			{
-				Namespaces:                    []string{"*"},
-				AllowIncomingLinks:            true,
-				AllowedOutgoingLinksHostnames: []string{"*"},
 			},
 		},
 		// Add background policies; policies that are not removed across
